pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["zavr1k <zavr35rus@gmail.com>"]
packages = [
    {include = "task_manager"},
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.2"
python-dotenv = "^0.17.1"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-bootstrap4 = "^3.0.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
coverage = "^5.5"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
